Move popover-confirm to new file
